feat: organized Search files
